Added persistence
