+++ Shopping List/Model Controller/ShoppingListController.swift
@@ -1,0 +1,36 @@
+//
+//  ShoppingListController.swift
+//  Shopping List
+//
+//  Created by Harmony Radley on 2/28/20.
+//  Copyright © 2020 Lambda School. All rights reserved.
+//
+
+import Foundation
+
+class ShoppingListController {
+    
+    var itemNames: [ShoppingItem] = [
+        ShoppingItem(name: "Apple"),
+        ShoppingItem(name: "Grapes"),
+        ShoppingItem(name: "Milk"),
+        ShoppingItem(name: "Muffin"),
+        ShoppingItem(name: "Popcorn"),
+        ShoppingItem(name: "Soda"),
+        ShoppingItem(name: "Strawberries"),
+    ]
+    
+ 
+    
+    
+    
+    
+    }
+    
+    
+    
+    
+    
+    
+    
+
