+++ Shopping List/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
