+++ Shopping List/Resources/Assets.xcassets/Grapes.imageset/grapes.png
