+++ Shopping List/Resources/Assets.xcassets/Soda.imageset/Contents.json
