+++ Shopping List/Resources/Assets.xcassets/Strawberries.imageset/strawberries.png
