+++ Shopping List/Storyboards/Base.lproj/Main.storyboard
@@ -31,11 +31,14 @@
                                     <rect key="frame" x="0.0" y="0.0" width="184" height="168"/>
                                     <autoresizingMask key="autoresizingMask"/>
                                     <subviews>
-                                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="grq-eG-dnw">
+                                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="Apple" translatesAutoresizingMaskIntoConstraints="NO" id="grq-eG-dnw">
                                             <rect key="frame" x="2" y="0.0" width="182" height="143"/>
                                         </imageView>
                                         <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="72c-bd-6g9">
                                             <rect key="frame" x="2" y="145" width="180" height="21"/>
+                                            <constraints>
+                                                <constraint firstAttribute="height" constant="21" id="DkY-I1-ope"/>
+                                            </constraints>
                                             <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                             <nil key="textColor"/>
                                             <nil key="highlightedColor"/>
@@ -84,12 +87,7 @@
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <textField opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="Enter your name:" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="MUQ-SM-lZr">
-                                <rect key="frame" x="57" y="209" width="283" height="34"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="14"/>
-                                <textInputTraits key="textInputTraits"/>
-                            </textField>
-                            <textField opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="Enter your address:" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="AYX-3h-Vp3">
-                                <rect key="frame" x="57" y="265" width="283" height="34"/>
+                                <rect key="frame" x="57" y="179" width="283" height="78"/>
                                 <fontDescription key="fontDescription" type="system" pointSize="14"/>
                                 <textInputTraits key="textInputTraits"/>
                             </textField>
@@ -109,18 +107,24 @@
                                     <action selector="sendOrder:" destination="azj-ze-F2n" eventType="touchUpInside" id="RYr-Sm-vrF"/>
                                 </connections>
                             </button>
+                            <textField opaque="NO" contentMode="scaleToFill" verticalHuggingPriority="251" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="Enter your address:" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="AYX-3h-Vp3">
+                                <rect key="frame" x="57" y="265" width="283" height="34"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="14"/>
+                                <textInputTraits key="textInputTraits"/>
+                            </textField>
                         </subviews>
                         <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
                         <constraints>
                             <constraint firstItem="kxl-M9-RST" firstAttribute="bottom" secondItem="AYX-3h-Vp3" secondAttribute="bottom" constant="563" id="2ZO-gv-dcj"/>
                             <constraint firstItem="8Hj-sg-ZUv" firstAttribute="centerX" secondItem="AYX-3h-Vp3" secondAttribute="centerX" id="5LD-5U-9nh"/>
                             <constraint firstItem="uxv-aO-fop" firstAttribute="leading" secondItem="kxl-M9-RST" secondAttribute="leading" constant="15" id="7dk-TQ-aGp"/>
-                            <constraint firstItem="AYX-3h-Vp3" firstAttribute="top" secondItem="MUQ-SM-lZr" secondAttribute="bottom" constant="22" id="GUI-fh-EBy"/>
+                            <constraint firstItem="AYX-3h-Vp3" firstAttribute="top" secondItem="MUQ-SM-lZr" secondAttribute="bottom" constant="8" id="GUI-fh-EBy"/>
+                            <constraint firstItem="MUQ-SM-lZr" firstAttribute="top" secondItem="uxv-aO-fop" secondAttribute="bottom" constant="-10" id="Kg1-eW-4we"/>
                             <constraint firstItem="uxv-aO-fop" firstAttribute="top" secondItem="kxl-M9-RST" secondAttribute="top" constant="22" id="SLI-gJ-Xzd"/>
                             <constraint firstItem="AYX-3h-Vp3" firstAttribute="trailing" secondItem="MUQ-SM-lZr" secondAttribute="trailing" id="Unt-v0-xOe"/>
                             <constraint firstItem="AYX-3h-Vp3" firstAttribute="leading" secondItem="kxl-M9-RST" secondAttribute="leading" constant="57" id="VVZ-7D-N0W"/>
                             <constraint firstItem="uxv-aO-fop" firstAttribute="centerX" secondItem="WuN-zA-uOY" secondAttribute="centerX" id="gBe-gs-neY"/>
-                            <constraint firstItem="AYX-3h-Vp3" firstAttribute="top" secondItem="MUQ-SM-lZr" secondAttribute="bottom" constant="22" id="igA-gt-AUM"/>
+                            <constraint firstItem="AYX-3h-Vp3" firstAttribute="top" secondItem="MUQ-SM-lZr" secondAttribute="bottom" constant="8" id="igA-gt-AUM"/>
                             <constraint firstItem="AYX-3h-Vp3" firstAttribute="leading" secondItem="MUQ-SM-lZr" secondAttribute="leading" id="phC-tM-xey"/>
                             <constraint firstItem="8Hj-sg-ZUv" firstAttribute="top" secondItem="AYX-3h-Vp3" secondAttribute="bottom" constant="26" id="v2j-Id-JRT"/>
                             <constraint firstItem="kxl-M9-RST" firstAttribute="trailing" secondItem="AYX-3h-Vp3" secondAttribute="trailing" constant="74" id="vbq-ZY-QhU"/>
@@ -128,6 +132,10 @@
                         <viewLayoutGuide key="safeArea" id="kxl-M9-RST"/>
                     </view>
                     <navigationItem key="navigationItem" id="0kp-JZ-9No"/>
+                    <connections>
+                        <outlet property="addressLabel" destination="AYX-3h-Vp3" id="C5A-2y-aFP"/>
+                        <outlet property="nameLabel" destination="MUQ-SM-lZr" id="MeL-xt-fOO"/>
+                    </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="iYK-9h-12D" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
@@ -152,4 +160,7 @@
             <point key="canvasLocation" x="562" y="5"/>
         </scene>
     </scenes>
+    <resources>
+        <image name="Apple" width="229" height="300"/>
+    </resources>
 </document>
